ml_pipeline_template/rag_pipeline/app/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
python-multipart
